+++ Utilities/Runtime/Extensions/UnityPoolExtensions.cs
@@ -5,12 +5,30 @@ namespace InfiniteCanvas.Utilities.Extensions
 {
     public static class UnityPoolExtensions
     {
-        public static List<T> GetList<T>(this T element) => ListPool<T>.Get();
+        public static List<T> GetList<T>(this T element)
+        {
+            var collection = ListPool<T>.Get();
+            collection.Add(element);
+            return collection;
+        }
 
-        public static void Release<T>(this List<T> collection) => ListPool<T>.Release(collection);
+        public static void Release<T>(this List<T> collection)
+        {
+            collection.Clear();
+            ListPool<T>.Release(collection);
+        }
 
-        public static HashSet<T> GetHashSet<T>(this T element) => HashSetPool<T>.Get();
+        public static HashSet<T> GetHashSet<T>(this T element)
+        {
+            var collection = HashSetPool<T>.Get();
+            collection.Add(element);
+            return collection;
+        }
 
-        public static void Release<T>(this HashSet<T> collection) => HashSetPool<T>.Release(collection);
+        public static void Release<T>(this HashSet<T> collection)
+        {
+            collection.Clear();
+            HashSetPool<T>.Release(collection);
+        }
     }
 }